Use nanosleep instead of usleep in tests
usleep(3) has been deprecated as of X/Open Issue 6 and its use on FreeBSD
(at least) results in a compiler warning when _XOPEN_SOURCE > 600. It doesn't
fail to link because FreeBSD's libc exposes the symbol when
_XOPEN_SOURCE <= 600 or _BSD_SOURCE (the equivalent to _GNU_SOURCE for
GNU extensions) is defined (the latter is the case here).

The sleep magnitudes have been adjusted by 1000 since usleep takes
values in microseconds and nanosleep takes values in seconds and
nanoseconds.

While here, fix the units for `conformance/interfaces/sem_post/8-1.c`
and `include/proc.h` to be long instead of unsigned long to align the
type with `<struct timespec>.tv_nsec`'s type, and make the value be in
seconds instead of milliseconds, calculating the number of iterations
appropriately. This is needed to avoid a long overflow issue on 32-bit
architectures, and helps simplify the code in the Linux case.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>