memcg_stress_test.sh: fix memory usage
Currently this test roughly calculates the maximum amount of
memory to use by barely looking up mem_free and swap_free from
/proc/meminfo.
Unfortunately it does not consider any special the systems
configured without a swap, so it ends, on those systems, trying
to use all the found mem_free while keeping only one mb
per-worker-thread as a safety net.
This is fine when using a 4k pagesize but it is NOT enough
on systems configured with a 64k pagesize, resulting in an
oom-party.
With this patch we now consider the configured page-size to
decide how many memory to set aside as a safety area.

Signed-off-by: Cristian Marussi <cristian.marussi@arm.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>