Test success with `mq_open` properly
Per the POSIX spec, testing for error from `mq_open` should be done like
so:

```
Upon successful completion, the function shall return a message queue
descriptor; otherwise, the function shall return (mqd_t)-1 and set errno
to indicate the error.
```

This worked on Linux because the return value was an integral value, as
opposed to some operating systems, like FreeBSD, where it is an opaque
pointer to a structure.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>