Address warnings
Fix warning related to mqdes being an int on Linux. Although it's
perfectly legal for mqdes to be initialized to NULL on FreeBSD, it is
not on Linux. Set the dummy variable to 0 to accomodate both
operating systems.

Fix a set but unused variable warning as well by casting the result of
the `mq_timedsend` call to void.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>