perf_event_open02: Add s390 specifics
Currently the testcase fails on s390 LPARs with enabled Counter Facility
as follows:

 perf_event_open02    0  TINFO  :  overall task clock: 43324672
 perf_event_open02    0  TINFO  :  hw sum: 801984103, task clock sum: 168166146
 perf_event_open02    0  TINFO  :  ratio: 3.881533
 perf_event_open02    1  TFAIL  :  perf_event_open02.c:346: test failed (ratio was greater than )

On s390 "buf.time_enabled" is always equal to "buf.time_running", therefore
count_hardware_counters() loops 1000 (MAX_CTRS) times and then returns 0.

The Counter facility on s390 provides numerous counters that are dedicated
to measure a specific activity, for example, instructions. Unlike other
architectures, it is not required to program a PMU and, thus, there are
no issues with counter constraints and multiplexing.

Therefore, adapt the test case to ignore multiplexing logic for s390 and
make the testcase work:

 perf_event_open02    0  TINFO  :  overall task clock: 61093214
 perf_event_open02    0  TINFO  :  hw sum: 200320299, task clock sum: 61091488
 perf_event_open02    0  TINFO  :  ratio: 0.999972
 perf_event_open02    1  TPASS  :  test passed

Signed-off-by: Michael Holzheu <holzheu@linux.ibm.com>
Reviewed-by: Hendrik Brueckner <brueckner@linux.ibm.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>