syscalls/mprotect04: align exec_func to 64 bytes
exec_func() is dummy/empty function. Try to align it so we don't
need to worry about copying 2 pages. But also check that compiler
aligned it and there's sufficient space between start of func_exec
and end of page.

This patch also removes copy_sz, which is now replaced with page_sz.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>