lib/tst_net.sh: add new TST_NET_MAX_PKT env variable
This is critical for performance tests, which compare virtual
devices with the real ones. Trigger the fragmentation but keep
the number of fragments at the minimum for those cases.

Also improve the test timing if MTU is not large.

The previous numbers can be restored with TST_NET_MAX_PKT
environment variable.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>