fcntl36: make sure write threads complete at least 1 loop
A failure has been reported while this test ran on a 1 CPU KVM guest
with 4.20 kernel:
  ...
  fcntl36.c:303: INFO: OFD r/w lock vs POSIX read lock
  fcntl36.c:360: FAIL: Unexpected data offset 3072 value 1

I wasn't able to reproduce it, to test the theory that
one of write threads was given time on CPU only after
loop_flag changed. Which means it wasn't able to complete
a single iteration.

This patch makes sure that write threads make at least
one iteration.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Xiong Zhou <xzhou@redhat.com>