syscalls/syncfs: Fix compilation
The lapi/syncfs.h calls tst_syscall() which in turn uses tst_brk and
TCONF, which is undefined until we include tst_test.h.

This commit just swaps the tst_test.h and lapi/syncfs.h includes so that
tst_test.h is included first, which fixed compilation on older
distributions. This is a quick fix, cleaner solution may be done later
on as a part of https://github.com/linux-test-project/ltp/issues/506.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>