lib/tst_numa: Add library for NUMA related syscalls
The library implements a nodemap, which is a map of nodes (array)
containing node ids for a subset of system nodes. Currently nodes can be
filtered based on memory requirements, which is needed for testing
memory related NUMA syscalls such as set_mempolicy() and membind() where
we usually need at least two memory nodes with free memory.

It also implements counters, so that we can precisely count, for a given
memory range, how much memory was allocated on each node in the map and
a few simple functions for mapping, faulting and unmapping memory.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
CC: Vlastimil Babka <vbabka@suse.cz>
CC: Michal Hocko <mhocko@kernel.org>
CC: Jan Stancek <jstancek@redhat.com>
Acked-by: Jan Stancek <jstancek@redhat.com>