lib/tst_checkpoint: Retry sleep on EINTR
Previously we haven't retried on EINTR, the main reason for that was
that all the tests using checkpoints haven't cared at all since signals
were not involved. However I think that retry on EINTR should be done by
default, since checkpoints returning from sleep on random signal
delivery are broken, and no test in the tree actually depends on such
behavior.

Currently tests for tgkill have emerged on ML where child needs to sleep
while parent sends signals to the child, this commit allows to use
checkpoints for these tests.

This commit implements the simpliest way how to retry, it's just a
simple loop over the futex() call that repeats on EINTR. The only side
effect is that timeout is restarted after the thread sleeping in futex()
got signal, since we use relative timeouts after all. However the
hypotetical worst case that can happen is that the test will timeout and
will end up being killed by the test library in a case that one test
would sleep on checkpoint while other would send signals in a loop.
Hoever chances for this arrangement are quite small.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Sumit Garg <sumit.garg@linaro.org>
Acked-by: Jan Stancek <jstancek@redhat.com>