setpriority01: Skip only PRIO_USER when unable to add test user
We don't need to skip all the tests just because we are unable to add
a test user. Not having a test user only affects PRIO_USER test case.
So just skip that one and continue running the rest of the tests when
useradd is not available or can't write to /etc/passwd (Eg: read-only
filesystem)

If useradd is present and it still fails for other reasons, then
consider it an error in the test preparation phase.

This also allows this test case to be built and run on Android.

Signed-off-by: Saravana Kannan <saravanak@google.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>