syscalls/fanotify01: add FAN_REPORT_FID test cases
For now, only check that event->fd is FAN_NOFD.
On old kernels the new test with result in TCONF.

tmpfs doesn't support FAN_REPORT_FID (it has zero fsid), so use a
blockdev filesystem. It is better to use a blockdev filesystem for
the rest of the test cases anyway to get more diversity in testing.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>