build: Move -Werror-implicit-function-declaration from make to build.sh
Previously it was passed only to Android build. Generally Werror flags
should be on for development but disabled for releases and production.
We don't have any configure flag stating development build, so using it
in build.sh should be sufficient as it can be used in both Travis CI
builds and local development.

+ respect CC variable in build.sh (for those lazy to pass it via -c flag)
+ change - to = after -Werror (same functionality, more common syntax).

Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>