hugetlb: limit hugepages according to MemAvailable
When the system's Hugepagesize is large(eg. 512MiB), the test may
break as cannot set nr_hugepages to default 128.

This patch continue to set 128 as default hugepages, but it checks
the memory available and hugepagesize registered. If default 128 is
larger than max available count, then reset the hugepages to 80% of
the max page count.

Without this patch:
  # ./hugemmap01
  tst_test.c:1096: INFO: Timeout per run is 0h 05m 00s
  mem.c:814: INFO: set nr_hugepages to 128
  mem.c:823: BROK: nr_hugepages = 54, but expect 128
  mem.c:814: INFO: set nr_hugepages to 0

With this patch:
  # ./hugemmap01  -s 100000000
  tst_test.c:1096: INFO: Timeout per run is 0h 05m 00s
  hugetlb.c:67: INFO: The hugepages beyond system available, here reset to 80% of the max page count 14502
  mem.c:814: INFO: set nr_hugepages to 11601
  hugemmap01.c:73: PASS: call succeeded
  mem.c:814: INFO: set nr_hugepages to 0

Signed-off-by: Xiao Liang <xiliang@redhat.com>
Signed-off-by: Li Wang <liwang@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>