Makefile: Fix target redefinition
There is no need to add a rule for the libs directory because there is a
generic rule for directory creation already. All we need to do is to
make the libs-all target depend on the directory path, which we do
already so that the directory is created before we attempt to execute
make there.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>