lib: add helpers to setup overlayfs mountpoint
Some tests are mounting overlayfs internally and run tests on it.
This mount can fail if the filesystem we are running on does not
support overlay mount upon it. For example, we are already running
tests on overlayfs or NFS, or CIFS. Test will report broken and
failure.

Fixing this by putting overlayfs dirs in a separated mountpoint,
using the same approach as in readahead02.

* define constraints of needed overlayfs dirs
* create_overlay_dirs() to create lower/upper/work dirs
* mount_overlay() to mount overlayfs
* SAFE_MOUNT_OVERLAY() macro to mount overlayfs safely, tst_brk TCONF
  if mount fail with errno ENODEV
* TST_MOUNT_OVERLAY() macro to mount overlayfs and return 0 if succeeds

Suggested-by: Petr Vorel <pvorel@suse.cz>
Suggested-by: Amir Goldstein <amir73il@gmail.com>
Reviewed-by: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: Murphy Zhou <xzhou@redhat.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>