scenario_groups/default: run the kernel_misc last
The block_dev test in kernel_misc will taint kernel with a warning flag,
this is an expected behaviour.

However, if you run the cve-2017-17053 after it, as the test is looking
for a warning / died flag for a tainted kernel, it will cause false
alarm with:
  tst_taint.c:88: BROK: Kernel is already tainted: 14849

The kernel_misc test was executed before the cve test in the default
test plan. Move the kernel_misc test to the end of the list to
workaround this issue.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>