syscalls/prctl07: Fix build.
Ideally this is last fix for prctl07.

Since I've changed the check_proc_capamb() to use macro definition
instead of raw integer value we depend on sys/capability.h being
included to compile the code.

So this commit changes:

The sys/capability.h is included under HAVE_SYS_CAPABILITY_H now and so
the function is ifdefed out if we don't have that header.

Also it looks like HAVE_LIBCAP is not defined unless libcap supports
cap_compare() which should be fixed. The check for libcap and
cap_compare() has to be uncoupled but that has to be done in a sepratate
patch.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>