syscalls/mprotect04: Remove itanium specific code
This fixes compilation error for itanium which was broken since 94f5be719.
While it'd be trivial to update that code to comply other arch code,
we agreed [2] to keep the fix as this arch is fading out and distros
officially supporting it are going to EOL in ~6 months.
Now, at least, the code compiles on itanium.

Fixes: 94f5be719 ("syscalls/mprotect04: align exec_func to 64 bytes")

[1] https://patchwork.ozlabs.org/patch/1137333/

Signed-off-by: Petr Vorel <pvorel@suse.cz>