build.sh: Respect user defined CFLAGS and LDFLAGS
Therefore it was needed to export them and not pass as a parameter.

It will be used for travis settings in next commit, but also can be
handy for using build.sh in local development.

Export variables instead of passing them to configure due issues with
whitespace in variable content.

Acked-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>