lib: Fix safe_mount() when exfat or ntfs are not using fuse
safe_mount() incorrectly assumes that when exfat or ntfs are attempted,
that they must be using fuse. However, ntfs and exfat can have proper
kernel drivers and thus mount successfully without the use of fuse. When
there *is* a proper kernel driver for ntfs or exfat, safe_mount()
attempts to use fuse and errors out.

Fix this by trying to mount the requested fs normally first, before
assuming that it needs fuse in the case of exfat and ntfs.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>