syscalls/madvise09: Workaround failures on newer kernels
This is a short term workaround that bumps the number of allocated pages
to 128 because which is needed since kernel commit:

1a61ab (mm: memcontrol: replace zone summing with lruvec_page_state())

After a discussion with kernel developers we concluded that in order to
avoid depending on kernel internal implementation details we have to
allocate a dozen of megabytes instead, but that would require a more
complex changes to the test. As we are getting close to a LTP release we
decided to put this workaround in first, then we can rewrite the test
after the release.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Shuang Qiu <shuang.qiu@oracle.com>
Acked-by: Li Wang <liwang@redhat.com>