fs/binfmt_misc02.sh: Fix local variable assignment for dash
Only the first one can be accepted if more than one string are assigned
to local valiable.  For example, only "This" is assigned to $string:
------------------------------------
tstring="This is test for extension"
...
local string=$tstring
------------------------------------

We add a pair of quotes to fix the issue now. Besides we can also fix
the issue by splitting declaration and assignment(e.g. local string &
string=$tstring).

See the explanation from:
https://wiki.ubuntu.com/DashAsBinSh#local

Fixes: #601
Reported-by: limingyu <limingyu@deepin.com>
Suggested-by: limingyu <limingyu@deepin.com>
Signed-off-by: Xiao Yang <ice_yangxiao@163.com>
Tested-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>