pan/ltp-pan.c: Add file parsing protection
This patch adds protection and print warning when the command part of
the tests is empty.

A segmentation fault occurs when given the wrong format in test cases
file for example:

    name  command line
    ----- -----------------
    dio10 diotest4 -b 65536    (O)
    dio10                      (X)
    dio10    diotest4 -b 65536 (X)

In this patch, ltp-pan will return empty collection when the given
testcase files are wrongly formatted and cause ltp-pan to abort. The
warning message will indicate the name of the error testcase.

Signed-off-by: Dylan Chung <dylan@andestech.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>