lapi/syscalls: Update syscall numbers
to sync with the kernel v5.6-rc3.

As this was performed in some sort of automated way, few extra changes
may be noticed while converting from kernel files to ltp, like a
different format of defining syscalls or removal of obsolete syscalls.

The diff generated against sh.in file is enormous and looks like someone
did a mistake while adding the syscall numbers there as they look to be
mostly incorrect from 220 to 393. They are fixed as a consequence of
this exercise.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
[ pvorel: various fixes for 32bit syscalls in 64bit arch and vice versa ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>