lib: tst_taint: Ignore WARN taint flag if already set
This commit changes the library so that it ignores the taint warn flag
if it was set prior to the test run. It turns out that the warn taint
flag is not well defined and could be easily set on a freshly booted
kernel for example when buggy BIOS is detected.

Other recent example is disabling ip forward on kvm guests:
  https://github.com/containers/libpod/issues/5815
  https://lore.kernel.org/netdev/a47b6a3b-c064-2f53-7cf6-d0d0720e9d99@redhat.com/

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Jan Stancek <jstancek@redhat.com>
Cc: Chang Yin <cyin@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>