kernel/syscall: remove the duplicate poll.h
this `poll.h` file is located at `testcases/kernel/syscalls/utils`
directory. Since we need to export the `testcases/kernel/syscalls/utils`
path as the reference directory for header files, so when we use
`#include <poll.h>` anywhere in the entire system, it will give priority
to referencing the `testcases/kernel/syscalls/utils/poll.h` instead of
the `<poll.h>` in the current system. As a result, this will lead to
some compilation errors.

In addition, some of the contents defined in `testcases/kernel/syscalls/utils/poll.h`
also conflicts with those defined in the system's `<poll.h>`.
For example, the definition of `struct pollfd {}` is in conflict.
Therefore, we can't solve the error of duplicate definition of `struct pollfd {}`
just by using `#include_next <poll.h>` directly in `testcases/kernel/syscalls/utils/poll.h`.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>